Move SimpleJsonProtocol to server.protocol
