+++ server/protocol/__init__.py
@@ -1,6 +1,7 @@
 from .gpgnet import GpgNetClientProtocol, GpgNetServerProtocol
 from .protocol import DisconnectedError, Protocol
 from .qdatastream import QDataStreamProtocol
+from .simple_json import SimpleJsonProtocol
 
 __all__ = (
     "DisconnectedError",
@@ -8,4 +9,5 @@ __all__ = (
     "GpgNetServerProtocol",
     "Protocol",
     "QDataStreamProtocol",
+    "SimpleJsonProtocol"
 )
